+++ packages/web-examples/vite-project/src/App.vue
@@ -1,10 +1,12 @@
 <script>
 import { ref, watch, onMounted, onUnmounted } from 'vue';
+
 import {
   preload,
   removeBackground,
   removeForeground,
   segmentForeground,
+  alphamask,
   applySegmentationMask
 } from '@imgly/background-removal';
 
@@ -12,10 +14,20 @@ export default {
   name: 'App',
   setup() {
     const images = [
-      'https://images.unsplash.com/photo-1686002359940-6a51b0d64f68?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2048&q=80',
-      'https://images.unsplash.com/photo-1590523278191-995cbcda646b?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=1080&fit=max&ixid=eyJhcHBfaWQiOjEyMDd9'
+      'https://images.unsplash.com/photo-1656408308602-05835d990fb1?q=80&w=3200&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+      'https://images.unsplash.com/photo-1686002359940-6a51b0d64f68?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1024&q=80',
+      'https://images.unsplash.com/photo-1590523278191-995cbcda646b?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=1080&fit=max&ixid=eyJhcHBfaWQiOjEyMDd9',
+      'https://images.unsplash.com/photo-1709248835088-03bb0946d6ab?q=80&w=3387&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D'
     ];
-    const randomImage = images[Math.floor(Math.random() * images.length)];
+
+    const url = new URL(window.location.href);
+    const params = new URLSearchParams(url.search);
+    const image = params.get('image');
+    const auto = params.get('auto') || false;
+    const randomImage = image
+      ? image
+      : images[Math.floor(Math.random() * images.length)];
+
     const imageUrl = ref(randomImage);
     const isRunning = ref(false);
     const seconds = ref(0);
@@ -27,6 +39,7 @@ export default {
     publicPath.pathname = '/js/';
     const config = {
       debug: false,
+      // debug: true,
       publicPath: publicPath.href,
       progress: (key, current, total) => {
         const [type, subtype] = key.split(':');
@@ -34,10 +47,17 @@ export default {
           0
         )}%`;
       },
+      // rescale: false,
+      rescale: true,
+      device: 'gpu',
+      // device: 'cpu',
+      model: 'isnet',
+      // model: 'isnet_fp16',
+      // model: 'isnet_quint8',
       output: {
         quality: 0.8,
-        // format: 'image/png'
-        format: 'image/jpeg'
+        format: 'image/png'
+        // format: 'image/jpeg'
         // format: 'image/webp'
         //format: 'image/x-rgba8'
         //format: 'image/x-alpha8'
@@ -96,11 +116,17 @@ export default {
     };
 
     const load = async () => {
+      const randomImage = image
+        ? image
+        : images[Math.floor(Math.random() * images.length)];
+
       isRunning.value = true;
       resetTimer();
 
       imageUrl.value = randomImage;
       const imageBlob = await removeBackground(randomImage, config);
+      // const imageBlob = await alphamask(randomImage, config)
+      // const maskBlob = await trimap(randomImage, config)
       // const imageBlob = await removeForeground(randomImage, config);
       // const imageBlob = await segmentForeground(randomImage, config);
       // const maskBlob = await segmentForeground(randomImage, {...config}); // use this format for maximum efficient
@@ -112,6 +138,8 @@ export default {
       isRunning.value = false;
       stopTimer();
     };
+
+    if (auto) load();
 
     return { imageUrl, isRunning, seconds, caption, load };
   }
@@ -141,9 +169,11 @@ export default {
   will-change: filter;
   transition: filter 300ms;
 }
+
 .logo:hover {
   filter: drop-shadow(0 0 2em #646cffaa);
 }
+
 .logo.vue:hover {
   filter: drop-shadow(0 0 2em #42b883aa);
 }