+++ packages/web/src/onnx.ts
@@ -1,57 +1,63 @@
 export { createOnnxSession, runOnnxSession };
 
 import ndarray, { NdArray } from 'ndarray';
-import * as ort from 'onnxruntime-web';
+import type ORT from 'onnxruntime-web';
+
+import * as ort_cpu from 'onnxruntime-web';
+import * as ort_gpu from 'onnxruntime-web/webgpu';
+
 import { loadAsUrl } from './resource';
-import { simd, threads } from './feature-detect';
+import * as feat from './features';
 import { Config } from './schema';
 
 async function createOnnxSession(model: any, config: Config) {
-  const capabilities = {
-    simd: await simd(),
-    threads: await threads(),
-    numThreads: navigator.hardwareConcurrency ?? 4,
-    // @ts-ignore
-    webgpu: navigator.gpu !== undefined
-  };
+  const useWebGPU = config.device === 'gpu';
+  const useThreads = await feat.threads();
+  const useSimd = feat.simd();
+  const proxyToWorker = config.proxyToWorker;
+  const executionProviders = [useWebGPU ? 'webgpu' : 'wasm'];
+  const ort = useWebGPU ? ort_gpu : ort_cpu;
+
   if (config.debug) {
-    console.debug('Capabilities:', capabilities);
+    console.debug('\tUsing Threads:', useThreads);
+    console.debug('\tUsing SIMD:', useSimd);
+    console.debug('\tUsing WebGPU:', useWebGPU);
+    console.debug('\tProxy to Worker:', proxyToWorker);
+
     ort.env.debug = true;
     ort.env.logLevel = 'verbose';
   }
 
-  ort.env.wasm.numThreads = capabilities.numThreads;
-  ort.env.wasm.simd = capabilities.simd;
-  ort.env.wasm.proxy = config.proxyToWorker;
+  ort.env.wasm.numThreads = feat.maxNumThreads();
+  ort.env.wasm.simd = feat.simd();
+  ort.env.wasm.proxy = proxyToWorker;
 
-  ort.env.wasm.wasmPaths = {
-    'ort-wasm-simd-threaded.wasm':
-      capabilities.simd && capabilities.threads
-        ? await loadAsUrl(
-            '/onnxruntime-web/ort-wasm-simd-threaded.wasm',
-            config
-          )
-        : undefined,
-    'ort-wasm-simd.wasm':
-      capabilities.simd && !capabilities.threads
-        ? await loadAsUrl('/onnxruntime-web/ort-wasm-simd.wasm', config)
-        : undefined,
-    'ort-wasm-threaded.wasm':
-      !capabilities.simd && capabilities.threads
-        ? await loadAsUrl('/onnxruntime-web/ort-wasm-threaded.wasm', config)
-        : undefined,
-    'ort-wasm.wasm':
-      !capabilities.simd && !capabilities.threads
-        ? await loadAsUrl('/onnxruntime-web/ort-wasm.wasm', config)
-        : undefined
+  const wasmPaths = {
+    'ort-wasm-simd-threaded.wasm': useThreads && useSimd,
+    'ort-wasm-simd.wasm': !useThreads && useSimd,
+    'ort-wasm-threaded.wasm': !useWebGPU && useThreads && !useSimd,
+    'ort-wasm.wasm': !useWebGPU && !useThreads && !useSimd
   };
+
+  const proxiedWasmPaths = {};
+  for (const [key, value] of Object.entries(wasmPaths)) {
+    if (value) {
+      const wasmPath =
+        useWebGPU && key.includes('simd')
+          ? `/onnxruntime-web/${key.replace('.wasm', '.jsep.wasm')}`
+          : `/onnxruntime-web/${key}`;
+      proxiedWasmPaths[key] = await loadAsUrl(wasmPath, config);
+    }
+  }
+
+  ort.env.wasm.wasmPaths = proxiedWasmPaths;
 
   if (config.debug) {
     console.debug('ort.env.wasm:', ort.env.wasm);
   }
 
-  const ort_config: ort.InferenceSession.SessionOptions = {
-    executionProviders: ['wasm'],
+  const ort_config: ORT.InferenceSession.SessionOptions = {
+    executionProviders: executionProviders,
     graphOptimizationLevel: 'all',
     executionMode: 'parallel',
     enableCpuMemArena: true
@@ -60,7 +66,7 @@ async function createOnnxSession(model: any, config: Config) {
   const session = await ort.InferenceSession.create(model, ort_config).catch(
     (e: any) => {
       throw new Error(
-        `Failed to create session: ${e}. Please check if the publicPath is set correctly.`
+        `Failed to create session: "${e}". Please check if the publicPath is set correctly.`
       );
     }
   );
@@ -70,8 +76,12 @@ async function createOnnxSession(model: any, config: Config) {
 async function runOnnxSession(
   session: any,
   inputs: [string, NdArray<Float32Array>][],
-  outputs: [string]
+  outputs: [string],
+  config: Config
 ) {
+  const useWebGPU = config.device === 'gpu';
+  const ort = useWebGPU ? ort_gpu : ort_cpu;
+
   const feeds: Record<string, any> = {};
   for (const [key, tensor] of inputs) {
     feeds[key] = new ort.Tensor(
@@ -82,9 +92,8 @@ async function runOnnxSession(
   }
   const outputData = await session.run(feeds, {});
   const outputKVPairs: NdArray<Float32Array>[] = [];
-
   for (const key of outputs) {
-    const output: ort.Tensor = outputData[key];
+    const output: ORT.Tensor = outputData[key];
     const shape: number[] = output.dims as number[];
     const data: Float32Array = output.data as Float32Array;
     const tensor = ndarray(data, shape);