+++ packages/web/src/schema.ts
@@ -2,6 +2,8 @@ export { Config, ConfigSchema, validateConfig };
 
 import { z } from 'zod';
 
+import * as feature from './features';
+
 import pkg from '../package.json';
 
 const ConfigSchema = z
@@ -22,6 +24,14 @@ const ConfigSchema = z
       .boolean()
       .default(false)
       .describe('Whether to enable debug logging.'),
+    rescale: z
+      .boolean()
+      .default(true)
+      .describe('Whether to rescale the image.'),
+    device: z
+      .enum(['cpu', 'gpu'])
+      .default('cpu')
+      .describe('The device to run the model on.'),
     proxyToWorker: z
       .boolean()
       .default(true)
@@ -36,7 +46,23 @@ const ConfigSchema = z
       .returns(z.void())
       .describe('Progress callback.')
       .optional(),
-    model: z.enum(['small', 'medium']).default('medium'),
+    model: z
+      .preprocess(
+        (val) => {
+          switch (val) {
+            case 'large':
+              return 'isnet';
+            case 'small':
+              return 'isnet_quint8';
+            case 'medium':
+              return 'isnet_fp16';
+            default:
+              return val;
+          }
+        },
+        z.enum(['isnet', 'isnet_fp16', 'isnet_quint8'])
+      )
+      .default('medium'),
     output: z
       .object({
         format: z
@@ -52,25 +78,49 @@ const ConfigSchema = z
       })
       .default({})
   })
-  .default({});
+  .default({})
+  .transform((config) => {
+    if (config.debug) console.log('Config:', config);
+    if (config.debug && !config.progress) {
+      config.progress =
+        config.progress ??
+        ((key, current, total) => {
+          console.debug(`Downloading ${key}: ${current} of ${total}`);
+        });
+
+      if (!crossOriginIsolated) {
+        if (config.debug)
+          console.debug(
+            'Cross-Origin-Isolated is not enabled. Performance will be degraded. Please see  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer.'
+          );
+      }
+    }
+
+    // always switch to gpu
+
+    if (config.device == 'gpu') {
+      if (!feature.webgpu()) {
+        if (config.debug)
+          console.debug('Switching to CPU for GPU not supported.');
+        config.device = 'cpu';
+      } else {
+        switch (config.model) {
+          case 'isnet':
+            break;
+          case 'isnet_fp16':
+          case 'isnet_quint8':
+            if (config.debug) console.debug('Switching to f32 model for GPU');
+            config.model = 'isnet';
+            break;
+        }
+      }
+    }
+
+    return config;
+  });
 
 type Config = z.infer<typeof ConfigSchema>;
 
 function validateConfig(configuration?: Config): Config {
-  const config = ConfigSchema.parse(configuration ?? {});
-  if (config.debug) console.log('Config:', config);
-  if (config.debug && !config.progress) {
-    config.progress =
-      config.progress ??
-      ((key, current, total) => {
-        console.debug(`Downloading ${key}: ${current} of ${total}`);
-      });
-
-    if (!crossOriginIsolated) {
-      console.debug(
-        'Cross-Origin-Isolated is not enabled. Performance will be degraded. Please see  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer.'
-      );
-    }
-  }
-  return config;
+  return ConfigSchema.parse(configuration ?? {});
 }