+++ packages/node/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- Added ThirdPartyLicenses.json Added
+- Added ThirdPartyLicenses.json
 
 ## [1.4.0]